NFV, XFV, XNFV + refactor tests
