+++ src/core/mod.rs
@@ -6,5 +6,5 @@ mod periodic;
 mod scheduled;
 
 pub use models::{DateLike, InvalidPaymentsError};
-pub use periodic::{fv, irr, mirr, nper, npv, pmt, pv};
-pub use scheduled::{xirr, xnpv, xfv};
+pub use periodic::{fv, irr, mirr, nfv, nper, npv, pmt, pv};
+pub use scheduled::{xfv, xirr, xnfv, xnpv};
