+++ tests/samples/xnfv.csv
@@ -1,0 +1,7 @@
+"2011-11-30", -100000
+"2012-03-15", -50000
+"2012-07-18", -2500
+"2012-11-30", 12500
+"2013-01-23", 37500
+"2013-04-30", 75000
+"2014-02-06", 90000
